requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
psycopg==3.3.0
psycopg-binary==3.3.0
python-dotenv==1.2.1
sqlparse==0.5.4
